Extract DrawingSyncCoordinator from CanvasView, rename tests to match
